requirements.txt
Flask
google-cloud-aiplatform
gunicorn